+++ Framework/iOS/ConvivaCore.framework/Headers/LivePass.h
@@ -103,4 +103,32 @@
 /// \deprecated Not used and to be phased out
 + (BOOL)ready DEPRECATED_ATTRIBUTE;
 
+/// @brief Set user preferences for data collection.
+/// \param userPrefs pass key-value pair as user preferences.
+
+/// Currently Supported keys: "idfa", "idfv".
+
+/// Expected Values:
+/// YES - Will collect id for mentioned key.
+/// NO - Will stop collecting id for mentioned key.
+/// If not been set, by default its considered as YES.
+///
+/// Sample usage:
+/// [LivePass  setUserPreferenceForDataCollection: [NSDictionary dictionaryWithObjectsAndKeys: @YES, @"idfa", @NO, @"idfv", nil]];
+
++ (void)setUserPreferenceForDataCollection:(NSDictionary *)userPrefs;
+
+/// @brief Set user preferences for data deletion.
+/// \param userPrefs pass key-value pair as user preferences.
+
+/// Currently Supported keys: "idfa", "idfv".
+
+/// Expected Values:
+/// YES - Will stop collecting id for mentioned key and delete previously collected data.
+/// NO - Will not have any effect.
+
+/// Sample usage:
+/// [LivePass  setUserPreferenceForDataDeletion: [NSDictionary dictionaryWithObjectsAndKeys: @NO, @"idfa", @YES, @"idfv", nil]];
++ (void)setUserPreferenceForDataDeletion:(NSDictionary *)userPrefs;
+
 @end